requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
boto3
botocore
pdfplumber
python-dotenv
streamlit
openpyxl
